Refactor: 폴더링 변경
