+++ app/src/main/java/com/sopt/now/util/base/BindingFragment.kt
@@ -1,4 +1,4 @@
-package com.sopt.now
+package com.sopt.now.util.base
 
 import android.os.Bundle
 import android.view.LayoutInflater
